# management-api-server/doc/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "This is a Restful service for operating Apache Cassandra.  You can find out more about the Management API on [Github](http://github.com/k8ssandra/management-api-for-apache-cassandra)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Management API for Apache Cassandra™",
    "version" : "0.1"
  },
  "paths" : {
...
    "/api/v0/ops/keyspace" : {
      "get" : {
        "operationId" : "listKeyspaces",
        "parameters" : [ {
          "in" : "query",
          "name" : "keyspaceName",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : [ "system_schema", "system", "system_auth", "system_distributed", "system_traces" ],
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            },
            "description" : "List of Keyspaces"
          }
        },
        "summary" : "List the keyspaces existing in the cluster"
      }
    },
    "/api/v0/ops/keyspace/alter" : {
      "post" : {
        "operationId" : "alterKeyspace",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateOrAlterKeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "Keyspace Replication Settings altered successfully"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "Altering Keyspace failed. Non-empty 'keyspace_name' must be provided"
              }
            },
            "description" : "Keyspace name or Replication Settings not provided"
          }
        },
        "summary" : "Alter the replication settings of an existing keyspace"
      }
    },
    "/api/v0/ops/keyspace/cleanup" : {
      "post" : {
        "operationId" : "cleanup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "d69d1d95-9348-4460-95d2-ae342870fade",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for keyspace cleanup process"
          }
        },
        "summary" : "Triggers the immediate cleanup of keys no longer belonging to a node. By default, clean all keyspaces. This operation is blocking and will return the executed job after finishing."
      }
    },
    "/api/v0/ops/keyspace/create" : {
      "post" : {
        "operationId" : "createKeyspace",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateOrAlterKeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "Keyspace created successfully"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "Keyspace creation failed. Non-empty 'keyspace_name' must be provided"
              }
            },
            "description" : "Keyspace name or Replication Settings not provided"
          }
        },
        "summary" : "Create a new keyspace with the given name and replication settings"
      }
    },
    "/api/v0/ops/keyspace/refresh" : {
      "post" : {
        "operationId" : "refresh",
        "parameters" : [ {
          "in" : "query",
          "name" : "keyspaceName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "table",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "SSTables loaded successfully"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "Must provide a keyspace name"
              }
            },
            "description" : "Keyspace name or Table name not provided"
          }
        },
        "summary" : "Load newly placed SSTables to the system without restart"
      }
    },
    "/api/v0/ops/keyspace/replication" : {
      "get" : {
        "operationId" : "replication",
        "parameters" : [ {
          "in" : "query",
          "name" : "keyspaceName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "class" : "org.apache.cassandra.locator.SimpleStrategy",
                  "replication_factor" : "2"
                }
              }
            },
            "description" : "Keyspace Replication Settings"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "Get keyspace replication failed. Non-empty 'keyspaceName' must be provided"
              }
            },
            "description" : "Keyspace name not provided"
          },
          "404" : {
            "content" : {
              "text/plain" : {
                "example" : "Get keyspace replication failed. Keyspace 'my_keyspace' does not exist."
              }
            },
            "description" : "Keyspace name not found"
          }
        },
        "summary" : "Get the replication settings of an existing keyspace"
      }
    },
...
    "/api/v1/ops/keyspace/cleanup" : {
      "post" : {
        "operationId" : "cleanup_v1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Cleanup not needed for tables 'system' or 'system_schema'"
          },
          "202" : {
            "content" : {
              "text/plain" : {
                "example" : "d69d1d95-9348-4460-95d2-ae342870fade",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for keyspace cleanup process"
          }
        },
        "summary" : "Triggers the immediate cleanup of keys no longer belonging to a node. By default, clean all keyspaces. This operation is asynchronous and returns immediately"
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateOrAlterKeyspaceRequest" : {
        "type" : "object",
        "properties" : {
          "keyspace_name" : {
            "type" : "string"
          },
          "replication_settings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReplicationSetting"
            }
          }
        },
        "required" : [ "keyspace_name", "replication_settings" ]
      },
...
      "KeyspaceRequest" : {
        "type" : "object",
        "properties" : {
          "jobs" : {
            "type" : "integer",
            "format" : "int32"
          },
          "keyspace_name" : {
            "type" : "string"
          },
          "tables" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "jobs", "keyspace_name", "tables" ]
      },
...
      "ReplicationSetting" : {
        "type" : "object",
        "properties" : {
          "dc_name" : {
            "type" : "string"
          },
          "replication_factor" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "dc_name", "replication_factor" ]
      },
...
    }
  }
}